TS/requirements.txt
asgiref==3.2.7
awsebcli==3.20.0
botocore==1.20.92
cement==2.8.2
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-import-export==2.2.0
django-js-asset==1.2.2
djangorestframework==3.11.0
et-xmlfile==1.0.1
future==0.16.0
idna==2.10
jdcal==1.4.1
jmespath==0.10.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
pathspec==0.5.9
Pillow==8.2.0
psycopg2==2.8.6
pypiwin32==223
python-dateutil==2.8.1
pytz==2020.1
pywin32==301
PyYAML==5.3.1
requests==2.25.1
semantic-version==2.8.5
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
termcolor==1.1.0
urllib3==1.25.11
wcwidth==0.1.9
xlrd==1.2.0
xlwt==1.3.0
